move rekordbox functions to seperate file
